backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
better-exceptions==0.3.3
alembic==1.13.1
pydantic-settings==2.4.0
sqlalchemy==1.4.50
pyjwt==2.8.0
google-generativeai==0.4.1